Convert IngestRecordsDynamoDB tests to use S3StateStore
