+++ java/ingest/ingest-runner/src/test/java/sleeper/ingest/IngestRecordsS3IT.java
@@ -19,10 +19,11 @@ import org.junit.jupiter.api.Test;
 
 import sleeper.core.record.Record;
 import sleeper.core.statestore.FileInfo;
+import sleeper.core.statestore.StateStore;
 import sleeper.ingest.testutils.AssertQuantiles;
-import sleeper.statestore.dynamodb.DynamoDBStateStore;
 
 import java.nio.file.Paths;
+import java.util.Collections;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -30,14 +31,14 @@ import static sleeper.ingest.testutils.IngestRecordsTestDataHelper.getRecords;
 import static sleeper.ingest.testutils.IngestRecordsTestDataHelper.getSketches;
 import static sleeper.ingest.testutils.IngestRecordsTestDataHelper.readRecordsFromParquetFile;
 
-public class IngestRecordsFromIteratorDynamoDBIT extends IngestRecordsDynamoDBITBase {
+public class IngestRecordsS3IT extends IngestRecordsS3ITBase {
     @Test
     public void shouldWriteRecordsCorrectly() throws Exception {
         // Given
-        DynamoDBStateStore stateStore = initialiseStateStore();
+        StateStore stateStore = initialiseStateStore();
 
         // When
-        long numWritten = ingestFromRecordIterator(stateStore, getRecords().iterator()).getRecordsWritten();
+        long numWritten = ingestRecords(stateStore, getRecords()).getRecordsWritten();
 
         // Then:
         //  - Check the correct number of records were written
@@ -50,16 +51,34 @@ public class IngestRecordsFromIteratorDynamoDBIT extends IngestRecordsDynamoDBIT
         assertThat(fileInfo.getPartitionId()).isEqualTo(stateStore.getAllPartitions().get(0).getId());
         //  - Read file and check it has correct records
         List<Record> readRecords = readRecordsFromParquetFile(fileInfo.getFilename(), schema);
-        assertThat(readRecords).containsExactly(getRecords().get(0), getRecords().get(1));
+        assertThat(readRecords).hasSize(2);
+        assertThat(readRecords.get(0)).isEqualTo(getRecords().get(0));
+        assertThat(readRecords.get(1)).isEqualTo(getRecords().get(1));
         //  - Local files should have been deleted
         assertThat(Paths.get(inputFolderName)).isEmptyDirectory();
         //  - Check quantiles sketches have been written and are correct (NB the sketches are stochastic so may not be identical)
-        AssertQuantiles.forSketch(getSketches(schema, activeFiles.get(0).getFilename()).getQuantilesSketch("key"))
+        AssertQuantiles.forSketch(getSketches(schema, fileInfo.getFilename()).getQuantilesSketch("key"))
                 .min(1L).max(3L)
                 .quantile(0.0, 1L).quantile(0.1, 1L)
                 .quantile(0.2, 1L).quantile(0.3, 1L)
                 .quantile(0.4, 1L).quantile(0.5, 3L)
                 .quantile(0.6, 3L).quantile(0.7, 3L)
                 .quantile(0.8, 3L).quantile(0.9, 3L).verify();
+    }
+
+    @Test
+    public void shouldWriteNoRecordsSuccessfully() throws Exception {
+        // Given
+        StateStore stateStore = initialiseStateStore();
+
+        // When
+        long numWritten = ingestRecords(stateStore, Collections.emptyList()).getRecordsWritten();
+
+        // Then:
+        //  - Check the correct number of records were written
+        assertThat(numWritten).isZero();
+        //  - Check StateStore has correct information
+        List<FileInfo> activeFiles = stateStore.getActiveFiles();
+        assertThat(activeFiles).isEmpty();
     }
 }
